langchain/test/requirements.txt
# python3 -m pip install -r ~/dev/study/langchain/requirements.txt
python-dotenv

langchain
langchain-openai
pymupdf
spacy # python3 -m spacy download ja_core_news_sm
tiktoken
chromadb